d/aws_service: Remove terraform-plugin-sdk implementation.
